update build version and improve config options
